+++ jest.config.base.js
@@ -59,7 +59,11 @@ const config = {
     // Enzyme setup file
     path.join(__dirname, 'client/shared/dev/enzymeSetup.js'),
   ],
-  setupFilesAfterEnv: [require.resolve('core-js/stable'), require.resolve('regenerator-runtime/runtime')],
+  setupFilesAfterEnv: [
+    require.resolve('core-js/stable'),
+    require.resolve('regenerator-runtime/runtime'),
+    require.resolve('@testing-library/jest-dom'),
+  ],
   globalSetup: path.join(__dirname, 'client/shared/dev/jestGlobalSetup.js'),
   snapshotSerializers: [path.join(__dirname, 'client/shared/dev/enzymeSerializer.js')],
 }